requirements.txt
pandas==1.3.3
mysql-connector-python==8.0.27
requests==2.25.1
beautifulsoup4==4.9.3
pymongo==3.12.0
google-api-python-client==2.38.0
Flask==2.0.1